Split all functions in main.cpp off into .h files in the include directory.
